backend/requirements.txt
flask
tensorflow
keras
opencv-python
google-genai
numpy
pandas
flask_cors
Pillow
python-dotenv